Feature/database configuration

+++ nistagram-backend/Nistagram/ChatMicroservice/Model/Message.cs
@@ -4,7 +4,7 @@ namespace ChatMicroservice.Model
 {
     public class Message
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Text { get; set; }
         public DateTime Date { get; set; }
         public Profile Profile { get; set; }
