+++ nistagram-backend/Nistagram/NotificationMicroservice/Model/Notification.cs
@@ -1,13 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace NotificationMicroservice.Model
 {
     public class Notification
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public DateTime Time { get; set; }
         public string Content { get; set; }
         public Profile Receiver { get; set; }
