+++ nistagram-backend/Nistagram/ProfileMicroservice/Model/User.cs
@@ -1,6 +1,8 @@
 ﻿using ProfileMicroservice.Model.Enum;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -8,7 +10,7 @@ namespace ProfileMicroservice.Model
 {
     public class User
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Username { get; set; }
         public string Password { get; set; }
         public string Email { get; set; }
